requirements.txt
# ============================================================
# Salamah Medical API — Unified requirements.txt
# Merged from Codebase A (mohammad) and Codebase B (shahd).
# Duplicates removed. Where versions differed the higher one is kept.
# onnxruntime-gpu (Codebase A) supersedes onnxruntime (Codebase B)
#   — it runs on both GPU and CPU environments.
# supabase-py is NOT included: the project uses SQLAlchemy exclusively.
# ============================================================

albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
bcrypt==4.0.1
boto3==1.42.71
botocore==1.42.71
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cloudinary==1.44.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
Cython==3.2.4
Deprecated==1.3.1
dnspython==2.8.0
easydict==1.13
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.135.1
flatbuffers==25.12.19
fonttools==4.62.1
greenlet==3.3.2
h11==0.16.0
idna==3.11
ImageIO==2.37.3
insightface==0.7.3
jmespath==1.1.0
joblib==1.5.3
kiwisolver==1.5.0
lazy-loader==0.5
limits==5.8.0
matplotlib==3.10.8
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
onnx==1.20.1
# onnxruntime-gpu covers both GPU and CPU; use plain onnxruntime if no GPU
onnxruntime-gpu==1.24.4
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
packaging==26.0
passlib==1.7.4
pgvector==0.4.2
pillow==12.1.1
prettytable==3.17.0
protobuf==7.34.0
psycopg2-binary==2.9.11
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
s3transfer==0.16.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
simsimd==6.5.16
six==1.17.0
slowapi==0.1.9
SQLAlchemy==2.0.48
starlette==0.52.1
stringzilla==4.6.0
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2026.3.3
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.41.0
wcwidth==0.6.0
wrapt==2.1.2
